architecture change
moved the test dot display function into the application source... the main component source should not contain application specific code.
